requirements.txt
flask
openai
pdfplumber
python-dotenv
